shop staff add stylist insert
