--- app/src/routes/error.ts
@@ -1,9 +1,0 @@
-export class InvalidRouteError extends Error {
-  constructor() {
-    super('Route not found')
-    this.code = 404
-    Object.setPrototypeOf(this, InvalidRouteError.prototype)
-  }
-
-  code: number
-}
