+++ app/src/routes/errors.ts
@@ -6,6 +6,7 @@ export const ServerError: MiddlewareErrorCode = 500
 export class MiddlewareError extends Error {
   constructor(message: string, code: MiddlewareErrorCode) {
     super(message)
+    this.name = 'MiddlewareError'
     this.code = code
     Object.setPrototypeOf(this, MiddlewareError.prototype)
   }
@@ -28,4 +29,15 @@ export class UnauthorizedError extends MiddlewareError {
   constructor() {
     super('User is unauthorized', Unauthorized)
   }
+}
+
+export class InvalidRouteError extends Error {
+  constructor() {
+    super('Route not found')
+    this.code = 404
+    this.name = 'InvalidRouteError'
+    Object.setPrototypeOf(this, InvalidRouteError.prototype)
+  }
+
+  code: number
 }
